Use flock to ensure exclusivity.
The pidof method currently employed does not ensure exclusivity as pidof
will not use argv[1] to find matching processes unless -x flag is
supplied (at least on Ubuntu Trusty and Wily, result may vary depending
on your distro and sysv version). Secondly, the current implementation
assumes the name of the process to be docker-gc, which may not be correct if
the user has named the binary something else.

This replaces the pidof implementation with one using flock.