Update cron.hourly documentation
I followed the previous documentation and created a `/etc/cron.hourly/docker-gc.sh` which was not actually run because it has the `.sh` extension. To make sure other users don't make the same mistake I explicitly mentioned `/etc/cron.hourly/docker-gc` and added a command to validate that the job will run.