Allow most recent n images to be preserved for a repo.
This allows MINIMUM_IMAGES_TO_SAVE=1 to be set to save the most recent
image for a repository. Any number of images can be saved, e.g. you can
save the most recent 10 images with MINIMUM_IMAGES_TO_SAVE=10.

By default MINIMUM_IMAGES_TO_SAVE is set to 0, which keeps the behavior
of docker-gc the same as before this commit.